Use single page app, the CallTo page is still in the source but not used.
Its faster to make a call and not have to go to another page to start typing a number.

When selecting a number in recent calls, copy to the textfield. Sometimes while scrolling it gets recognized as selected and starts a call.
Add call button to textfield.
Use systems Icons.